+++ frontend/src/pages/sales/SalesReturn.tsx
@@ -74,7 +74,7 @@ const SalesReturn = () => {
   }, [saleDetails]);
 
   const handleReturnQtyChange = (index: number, val: string) => {
-    const qty = val === '' ? '' : parseInt(val) || 0;
+    const qty = val === '' ? '' : parseFloat(val) || 0;
     const newItems = [...returnItems];
     // prevent returning more than sold
     if (typeof qty === 'number' && qty > newItems[index].quantity) return;
